HTML / CSS files needed for website
I zipped it because it had too many files. This is an incomplete website with modifications and customizations that can be used together with the javascript. To open, just unzip, and open index.html, additional pages can be found in /pages. Feel free to make changes to any of the individual pages. This link may help: https://fortawesome.github.io/Font-Awesome/cheatsheet/