Morse translator
Zhuoen - When you add this to the page keep these things in mind:

The HTML I have there is essential to run the page (not the <p> tags, but they are useful descriptors). I have the output logged to console (see line 81), but you can choose where you put the output (through dynamic HTML).